Adds support functions and types to utils folder
