+++ react/utils/bindings/typings.d.ts
@@ -1,0 +1,5 @@
+export interface DropdownChangeInput {
+  target: {
+    value: string
+  }
+}
